Added the ability to custom which topics and sections that should be shown. See the different declares of show_help(...).
All sections and topics are now saved to be referenced through show_help(...), even though they are not referenced from the toplevel.

The configuration file is now read and parsed when the help manager is created,
which should prevent unnecessary delays when opening the browser.

Links are now only shown as links if the topic the refer to is shown. Added a force= attribute to the reference tags which make them always show the link, even though the topic is not shown (although trying to follow the link will fail).


git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@2452 75b3775d-3801-0410-9e12-e2e43121521d