Revert all indexed/greyscale back to RGB.
DO NOT run agressive png optimizations on trunk anymore. It corrupts
some images (see rev. 10243 as an example of this, all void tiles lost
their alpha gradient in that commit). It's also very inconvenient for
the artists. Rather run opt. before a release, but beware of
corruption like described above.


git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@10481 75b3775d-3801-0410-9e12-e2e43121521d