First step towards making the SVN revision available for display; teach the build machinery to include a SVNREV definition in compiles, and use it to initialize a 'svnrev' constant in game_config.cpp.
This needs two more pieces.  First, we need to coax SVN into touching a 
revision-stamp file on every commit, and make that an ancestor of 
game_config.cpp.  (As it is now, the 'svnrev' constant will only 
be refreshed when game_config.cpp is recompiled.)

Second, we actually need to find the right place to display this.
Or we need to paste the svnrev constant onto the end of the
version string, but only when the former ends in 'svn'. 


git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@16923 75b3775d-3801-0410-9e12-e2e43121521d