Major prose revisions for An Orcish Incursion.
As originally written by Josh Parsons, this
campaign was not set in any particular time
or place.  I've changed it to a few years after 
the arrival of orcs on the Great Continent, 
and pinned it to the northern marches of Wesmere.
Erlornas and his troops were professional guards,
but I've changed them to a scratch force of civilians 
raused by a local noble and gradually militarizing.

Thus, this becomes a story of how elves learned
the nature of orcs and how to cope.  I've also
added some development of their ideas about humans,

The plot is still linear, but there is a lot more 
forward and backward reference in the story arc
now, so it reads much less like a series of unmotivated
and disconnected incidents. 


git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@20856 75b3775d-3801-0410-9e12-e2e43121521d