Address bug #8601.  The fundamental problem here is that the meaning of the hourglass field changes dynamically according to whether you're in SP or MP, while themes are static; thus, whatever we put in as a static tooltip in the theme is going to be wrong a lot of the time.
The fast solution is to rip this tooltip out, which is what I've
done.  With some C++ klugery analogous to what's done to change the
End Turn button label in linger mode we could do something clever
here, but I don't think the small gain would be worth the extra complexity.

Anyone who disagrees with me is welcome to file a FR, or (better yet) 
have at the C++ themselves.


git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@21773 75b3775d-3801-0410-9e12-e2e43121521d