correction to r27544 - apply traits and modifications to attacks in unit::read()
note: this code needs an overhaul! The "real" max hitpoints, etc should be stored in unit::write and WML changes to these should be allowed to override the calculated values in unit::read. Also the _b_ variables in unit.cpp/hpp should be removed and reference the unit_type directly when a "base" value is needed. 

git-svn-id: svn+ssh://svn.gna.org/svn/wesnoth/trunk@27577 75b3775d-3801-0410-9e12-e2e43121521d