Initial implementation of the scaling image.
We now use a 400x400 image (max size 500x500) and scale that down to a
'nice' size for the current window size. The scale factor might need
some tuning.
Kitty also prefers to see a 400x400 image scaled down and evaluate the
quality, this might means we only need a 400x400 image but also that we
still need a 300x300 and maybe a 200x200 image.
It's still not possible to manipulate the new portraits from the WML but
that's postponed until we know how much images will be used.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@30593 75b3775d-3801-0410-9e12-e2e43121521d