Let the window builder return a pointer.
The current design of twindow (and some other widget classes) have the
problem that they shouldn't be copied. The code doesn't copy the objects
as long as the compiler does a return value optimization, which
indicates a bad class design. By turning the return value into a pointer
the copying is no longer needed.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@31727 75b3775d-3801-0410-9e12-e2e43121521d