Add better draw support for invisible items.
Items in a scrollbar container which are not visible are no longer
redrawn even it they're dirty. (This was observed if a language which
was selected and scrolled out of sight still showed up when deselected.)
Items that are partly visible are still fully redrawn, this will be
addressed later.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@31854 75b3775d-3801-0410-9e12-e2e43121521d