First step in the unit profile hacks.
The time until 1.6 is too short to write and convert all mainline stuff
to use the new portrait so wrote a hack. Now the engine looks at the
profile key again and tries to find the same image in the transparent
subdirectory, if found that image is used instead. The following issues
still need to be resolved:
- images with the size 205x205 shouldn't be scaled.
- right size only images are not implemented yet.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@32188 75b3775d-3801-0410-9e12-e2e43121521d