Two more fixes discovered while checking translations:
"Edition" -> "Editing": The former was used incorrectly instead of
the active verb.

"the godly city of Tath" -> "the goodly city of Tath": some non-native speaker
probably thought "goodly" was a typo.  It isn't, just an archaic form of the
adjective "good".

In both cases it is not really necessary to change the actual translations, 
but the string IDs containing these should have the spelling fixed as .po
files come back in for integration.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@32540 75b3775d-3801-0410-9e12-e2e43121521d