Fix the mouse focus after a missed event.
When clicking fast on a button the engine might miss not have the proper
focus (the previous commit should fix that but it seems you can click
fast enough to get the mouse down event in the queue before the move
event). If that happens set the focus in the mouse down event. This
fixes bug #12927.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@32884 75b3775d-3801-0410-9e12-e2e43121521d