Properly reset the scrollbar mode when resizing.
Before after the first sizing and the scrollbar not needed it was set to
a disabled mode, which meant the scrollbar would no longer be shown when
needed after resizing. Fixes bug #13018.

Another unrelated bug was discovered while fixing this issue, the side
panel gets wrongly overdrawn. This due to the undrawing, which will be
fixed post 1.6.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@33252 75b3775d-3801-0410-9e12-e2e43121521d