Fix continuity issues created by Chantal becoming personally important to Delfador - happens in Delfador's Memoirs, affects Heir top the Throne.
This patch bumps Chantal from Druid to Shyde, and makes her death in HttT a
loss condition - though one that is highly unlikely to trigger given the
usual flow of battle in the scenario and the hitpoints she has as an L3.

This patch also adds dialog to both Elves Beseiged and Siege of Elensefar
that improves contnuity with LoW, with DM, and internally in HttT itself.

Finally, Chantal's name becomes a global spelling exception. 


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@34575 75b3775d-3801-0410-9e12-e2e43121521d