Terrain types can now have description attributes.  If present, these override the name attribute for the "outside" portion of the human-readable terrain description, e.g.  the part not in parentheses.
The main use for this is editor palette tooltips; they now say, for example,
"Summer Deciduous Forest (Forest)".  I've also wired things so that the normal
in-game terrain report uses description rather than name when it's available, 
as a demonstration (that's the diff to generate.report.cpp) but it might be
that level of detail is not actually desirable in the game itself.  


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@34933 75b3775d-3801-0410-9e12-e2e43121521d