Add a new initial implementation for lexical_cast.
The code is only used in the unit test at the moment. It would be nice if
others can test whether the code works on all supported platforms since it
makes more use of templates. If there are no reports of issues, I'll start to
use this code to replace the old lexical cast.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@35418 75b3775d-3801-0410-9e12-e2e43121521d