Change the handling of double clicks.
The engine now always tries to generate a double click, if a double
click is generated it looks whether a widget wants double clicks when
building the event chain. Depending whether or not it supports a double
click it tries to send a double or single click.
After this descission it looks whether a handler is installed, but only
test the click or double click.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@39281 75b3775d-3801-0410-9e12-e2e43121521d