Add an overload for set_row_shown.
Hiding one row at a time in the game_load dialog, when the filter
resulted in an empty set was slow. Every time a row was hidden the next
row was selected and its callback called. Directly after that the newly
selected row was hidden as well... The speedup is quite noticable.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40312 75b3775d-3801-0410-9e12-e2e43121521d