Non-human null controllers are no longer set to ai.
This avoids dummy sides to get a turn, like the statues in 'Caves of the
Basilisk'.

Regarding compatibility: I did some minimal testing with a version with
and without this patch and it seemed to work properly. Since we're not
in compatibility freeze I didn't do more extensive testing. (This fix
also allows me to properly fix another bug [1], to be committed later).

[1] http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=563310

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40514 75b3775d-3801-0410-9e12-e2e43121521d