Add experimental option.  This defaults to 'yes' for odd-numbered (unstable) versions, 'no' for even-numbered (unstable) versions. It controls an EXPERIMENTAL preprocessor symbol, defined only when experimental is 'yes', and an EXPERIMENTAL automake conditional.
Thio is meant as a guard option for experimental, developer-only 
code that should be disabled in stable versions.  The immediate
application is fendrin's new pathfinding/tunnel code, which we want
to carry in the main codebase for the remainder of 1.7 but be automatically
disabled in 1.8 betas and 1.8 itself.

Implemented for scons and auttools but not yet cmake.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@40867 75b3775d-3801-0410-9e12-e2e43121521d