Fix bug #15429
r40928 forces the Neutral alignment when it isn't overridden by
SingleUnitWML, assuming the empty/missing alignment specification to be
"invalid". The logic before that commit was also broken, so now we make
sure that empty/missing alignment specifications are ignored (they seem
to be automatically fixed to inherit the unit type's alignment later in
the code).

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@41285 75b3775d-3801-0410-9e12-e2e43121521d