Fix a crash when using the scrollwheel.
A while ago I allowed the container area to be larger as the content.
This change allowed some bounds checking problems, allowing the scroll
outside the visible area when all items were visible. Now test whether
all items are visible, if so only allow scrolling to the beginning of
the list.

Fixes bug #15156.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@41498 75b3775d-3801-0410-9e12-e2e43121521d