Now the top item in the stacked widget is selected.
Also change the order of the layers in the new title screen to make sure
the proper layer is still selected. The wanted side effect is that the
logo is now drawn at the bottom.

This changes the behaviour or widgets with multiple selections, but the
stacked widget is the only one, so no problems in other places.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@41989 75b3775d-3801-0410-9e12-e2e43121521d