Reimplemented the tooltip event code.
The code now lets the widget show its own tooltip instead of letting the
distributor figure it all out.

The commit also adds a new event type class; the message. A message goes
from a widget to the window (instead of the other way around). The message
class is used to let the widget ask the window to really show the tooltip.
The reason for doing so is to both decouple who needs to show the tooltip
(no reason why the window is responsible) and to add the new message class.
(I've some usage ideas for the class, but nothing to commit yet.)


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@48490 75b3775d-3801-0410-9e12-e2e43121521d