removed Dev-C++ project files
Nobody has been maintaining them since more than 2 years.
It'll be a ton of work to bring them up to date (so it's probably a better choice
to convert the CodeBlocks files instead). The Dev-C++ IDE itsself is a dead
project since 2005. Nobody will think there's need to update these files anymore
when adding source files. The compiler used was MinGw, so CodeBlocks is basically
a replacement.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@49866 75b3775d-3801-0410-9e12-e2e43121521d