Fixed compilation on all Debian architectures.
The #undef was introduced in r47336, but the commit offers no reason
_why_ the macro was undefined. This patch has been tested by Rhonda on
all Debian architectures and there it caused no build problem, whereas
the r47336 caused several architectures failing to build.

The reason the issue was spotted much later is that before Debian
compiled with the ANA network. It switched to the older network code
while changing the build system from autotools to cmake, so the latter
was thought to be the cause.

Fixes Debian bug #636193 and reverts r47336.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51130 75b3775d-3801-0410-9e12-e2e43121521d