Fix for bug #18745 It is now impossible to jump to ability topics from th esidebar in most translations
Changed all vectors of abillity_tooltip objects to be vectors of t_strings.  Previously, sometimes they passed std:string and sometimes t_strings. This meant that the help index was a t_string->translated string ->t_string ->base_str(), where the base string was actually a translated string.  The tooltip was passing t_string->translated_string as index which, although inconsitent worked.

Now when the tooltip is created it uses the translated string as the visible string and the base_str() as the index into the help, as intended but not done in the original code.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51335 75b3775d-3801-0410-9e12-e2e43121521d