Add a work-round for bug #18534.
The destructor of the class restored the moment of the units. When a
unit no longer exists it overwrites random memory. The issue has been
found by valgrind.

anonymissimus tested the work-around since he was able to reproduce the
issue easily, while on my system it went well for a long time.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51485 75b3775d-3801-0410-9e12-e2e43121521d