set playsingle_controller::end_turn_ to false in linger mode (fix for bug #16821)
The problem was that this variable gets usually set to false during
the start of the side turns. If check_victory() ends the level between
the side turns, this initialization is missed so linger mode thinks
it has already been ended.
Other call cases of linger() should call the function with end_turn_==false
already. That is, I think they already do, since if not these would have
fallen into this bug category as well.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51600 75b3775d-3801-0410-9e12-e2e43121521d