Sort-of revert r50745
The only thing this commit did was mess around with dangled pointer
access timings instead of solving the original UB that it was supposed
to fix (see bug #18495). There's no problem with accessing theme UI
buttons before they are initialized since they will be NULL and the
replay controller already made sure to not do silly crap with NULL
pointers (except for the case spotted in r51823).

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@51825 75b3775d-3801-0410-9e12-e2e43121521d