Fixed the translation of some tutorial messages.
The second part of the narrator messages were not translated, not sure
why. This patch is a work-around not a proper fix. Once 1.10 is in its
own branch a proper fix can be committed. That fix will change
translatable strings, and is the reason for a work-around in the trunk.

Fixes bug #19316.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@52785 75b3775d-3801-0410-9e12-e2e43121521d