Remove a backwards compat hack left over from r44301 and 44301.
The hack causes problems now, since I want to load helper.lua previously
already and the comment says that it should be removed in 1.11.
I *think* the reason for the hack was to avoid loading wml-tags.lua twice
in case that a replay and/or (perhaps) savegame is loaded which was
created with a BfW version previously to r44301, since those replays/saves
seemed to have recorded the commands for loading the core lua files, which
is no longer the case since back then. Those replays/saves are no
longer relevant, so everything hints that I can safely remove this.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@52976 75b3775d-3801-0410-9e12-e2e43121521d