introduce inline SSF support in [objectives]
No deprecation needed since it defaulted to the 'default
objectives' already. They are used for a side in case that that side
doesn't have side specific objctives set. Use a SSF which matches
no sides or all sides to set the default objectives, if a subset
of all sides match, the side specific objectives for these are set.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53463 75b3775d-3801-0410-9e12-e2e43121521d