Improve the multiply instruction.
It now keeps more resolution and is less prone to overflows (unless of
course the result overflows). The code is aimed at an ARM v7 so might be
slow on platforms that do not support an instruction for:
32-bit signed * 32-bit signed = 64-bit signed

The alternative was to write code like for the division, but that would
be more code and slower on platforms with fast 64-bit multiplication
result.

The output of the titlescreen image now looks less dark as before this
modification.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53505 75b3775d-3801-0410-9e12-e2e43121521d