Replace non-standard Underground ToD schedule picture (125x38) with an upscaled version of the correct size (125x39), provided in bug #19415
This solution seems sensible for reasons stated in the bug report. I
normally dislike upscaling graphics, but in this case the original is
crappy enough (it even appears to have JPEG artifacts?) and the
difference barely noticeable to the human eye.

Since this kind of changes can apparently break add-ons nowadays, this
probably should _not_ be backported to 1.10.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53568 75b3775d-3801-0410-9e12-e2e43121521d