addon/mg: Force version_info cache refresh after installing add-ons
The cache obviously becomes goes stale after installing/upgrading any
amount of add-ons. It's normally refreshed in the same steps as all
other caches (WML, graphics, etc.), but when we are in the manager UI
loop that code doesn't get a chance to run, so we must reset this
particular cache before continuing. This makes add-on statuses work
correctly.

In general, the version info cache was a terrible idea, but perhaps its
underpinnings could be adapted for add-on dependency/conflict resolution
in the future.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53582 75b3775d-3801-0410-9e12-e2e43121521d