set MSVC's intermediate output directories correctly
I finally managed to solve the linker warning spam about missing debug
information for liblua. Apparently the file MSVC was looking for was
overwritten by the one for wesnoth wich has the same name. This sets
an individual directory to write it to as it's already done for
wesnothlib. Also, this finally enables me to debug deep down into the
lua upstream source. :)
Also some other mostly warning related adjustments and fixes.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53682 75b3775d-3801-0410-9e12-e2e43121521d