Fix binding to a temporary.
The result of std::string::c_str() of a temporary std::string was stored
in an object to be used later. Changed the member to a std::string to
solve the issue.

NOTE since the code involved contains serveral OS specific #ifdef's
not all changes are tested.

Issue found by cppcheck; it choked on the code.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53854 75b3775d-3801-0410-9e12-e2e43121521d