Make leader_image paths in the saved games index binary path-independent (part of bug #18683)
This makes it possible to store in the save_index image paths that can
be used in the Load Game dialog even when displaying units whose
graphics are in binary paths not currently loaded. This alleviates the
effect of bug #19410 on the whole mechanism, but not completely.

When the save_index is missing, this won't work properly; we'll need to
add the leader_image attribute to the saved game config itself later.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@53932 75b3775d-3801-0410-9e12-e2e43121521d