Created basic implementation of the viewport.
The class can now hold a widget and forward calls to it. The class is
not yet complete but it looks much better now.

Also allows to switch between the naked pane and the pane in a viewport
with a #if 0.

The code is used to experiment with a different approach of the
implementation of a listbox.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54100 75b3775d-3801-0410-9e12-e2e43121521d