Implement the expand and collapse button.
The items  in the adon list can now be expanded and collapsed, by
default the description is collapsed.

The expanding and collapsing causes a resize event not handled by the
pane causing a huge slowdown. This will needs some love.

The code is used to experiment with a different approach of the
implementation of a listbox.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54136 75b3775d-3801-0410-9e12-e2e43121521d