Add a debug and test code.
In order to test with formatting growing listboxes I need a growing
listbox. When showing the addon dialogue only one is shown, the others
are loaded with the load button. The code is a bit hacky, but hidden
under the --new-widgets switch.

The code is used to experiment with a different approach of the
implementation of a listbox.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54172 75b3775d-3801-0410-9e12-e2e43121521d