Fix the formatting for a growing pane.
The code seems to work for this specific case but needs more testing.

This commit finishes the proof-of-concept prototype it seems the big
hurdles in the listbox have been tackled. Now need to focus on getting a
real widget out of this prototype.

The code is used to experiment with a different approach of the
implementation of a listbox.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54173 75b3775d-3801-0410-9e12-e2e43121521d