addon/mg: New add-on installation prompt for add-ons with VCS/.pbl files
Until now there was only a single, verbose prompt for the .pbl case.
Since people may use a VCS to contribute to an add-on without having a
.pbl file, that case is also now covered by this commit.

There's a single prompt with a list of sensible content that might be
"permanently lost" on installation. The list may contain VCS or .pbl
files for now, but perhaps at some point in the future someone might
want to make it into a full list of extra files or something.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54272 75b3775d-3801-0410-9e12-e2e43121521d