Add a new instance dummy widget.
This widget is used to postpone instantiation of a widget. This will be used
in the new matrix widget where the definition needs a widget which will only
be available when the matrix widget is instantiated in a window.

The code is used to experiment with a different approach of the
implementation of a listbox.


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54322 75b3775d-3801-0410-9e12-e2e43121521d