addon/mg: Greatly simplify UI logic when no add-ons can be displayed
We now display the add-ons manager with an amount of controls
hidden/disabled and a line of text explaining the situation in a few
words.

This was necessary to reduce complexity since there are now more
situations to handle because of the add-on type filtering options. It
should look better this way, anyway, and should even be less annoying
for the users since the dialog no longer resets the view mode on its
own.

git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54351 75b3775d-3801-0410-9e12-e2e43121521d