Refectored move_unit()
This more object-oriented approach should be more robust in the face of
whatever future changes might come along. Plus, the code might be easier
to follow since the logic is broken into smaller pieces.
(There is some more streamlining to come. At this stage, very little was
changed with regards to the animation support that is required by the
enter/exit hex events.)


git-svn-id: http://svn.gna.org/svn/wesnoth/trunk@54467 75b3775d-3801-0410-9e12-e2e43121521d